feat: site optimization
suggested github actions flow
example: https://github.com/filipesilva/ng-github-actions
reversed polyfill and utilities function for re-use
