+++ .github/workflows/deploy.yml
@@ -28,11 +28,11 @@ jobs:
       - name: Run linting
         run: npm run lint
 
-      - name: Run unit tests
-        run: npm run test:ci
-
       - name: Build
         run: npm run build:misc && ng build --configuration=staging
+
+      - name: Run unit tests
+        run: npm run test:ci
 
       - name: Configure AWS credentials
         uses: aws-actions/configure-aws-credentials@v1
