+++ src/polyfills.ts
@@ -57,24 +57,9 @@
  */
 // global window object needed for awsamplify as per https://github.com/aws-amplify/amplify-js/issues/678
 (window as any).global = window;
-/**
- * https://developer.mozilla.org/en-US/docs/Web/HTML/Preloading_content
- * https://support.mozilla.org/en-US/questions/1248200
- * Not supported in firefox, safari.
- */
-(() => {
-  const doc = document as any;
-  const styleSheetLink = doc.createElement("link");
-  const supportsPreload = styleSheetLink.relList.supports("preload");
-  if (!supportsPreload) {
-    const preLoads = doc.head.querySelectorAll(
-      'link[rel="preload"][as="style"]'
-    );
-    preLoads.forEach((element: HTMLLinkElement) => {
-      element.setAttribute("rel", "stylesheet");
-    });
-  }
-})();
+// hee-shared application utility function
+import { PreloadStyleSheet } from "hee-shared";
+PreloadStyleSheet();
 import "zone.js/dist/zone"; // Included with Angular CLI.
 
 /***************************************************************************************************
